+++ tests/testthat/test-prepInputs.R
@@ -94,6 +94,7 @@ test_that("prepInputs doesn't work", {
   expect_true(is(shpEcozoneSm, "SpatialPolygons"))
   expect_identical(extent(shpEcozoneSm), extent(StudyArea))
 
+  unlink(dirname(ecozoneFilename), recursive = TRUE)
   # Test useCache = FALSE -- doesn't error and has no "loading from cache" or "loading from memoised"
   mess <- capture_messages(shpEcozoneSm <- Cache(
     prepInputs,
@@ -161,7 +162,7 @@ test_that("prepInputs doesn't work", {
   reproducible::clearCache(userTags = "prepInputs", ask = FALSE)
   # previously, this would cause an error because prepInputs file is gone b/c of previous
   #  line, but postProcess is still in a Cache recovery situation, to same file, which is
-  #  not there. Now should be no error
+  #  not there. Now should be no error.
   mess <- capture_messages(LCC2005_2 <- Cache(
     prepInputs,
     url = url,
@@ -1380,8 +1381,8 @@ test_that("options inputPaths", {
                         dlFun = getDataFn, name = "GADM", country = "LUX", level = 0,
                         path = tmpCache)
   })
-  expect_true(sum(grepl("Hardlinked version of file created", mess1))==1) # used a linked version
-  expect_true(sum(grepl(basename(tmpdir2), mess1))==1) # it is now in tmpdir2, i.e., the destinationPath
+  expect_true(sum(grepl("Hardlinked version of file created", mess1)) == 1) # used a linked version
+  expect_true(sum(grepl(basename(tmpdir2), mess1)) == 1) # it is now in tmpdir2, i.e., the destinationPath
 
   # Have file in destinationPath, not in inputPath
   unlink(file.path(tmpdir, theFile))
@@ -1398,7 +1399,6 @@ test_that("options inputPaths", {
 
 })
 
-
 test_that("writeOutputs saves factor rasters with .grd class to preserve levels", {
   skip_on_cran()
 
@@ -1406,14 +1406,17 @@ test_that("writeOutputs saves factor rasters with .grd class to preserve levels"
   on.exit({
     testOnExit(testInitOut)
   }, add = TRUE)
-  a <- raster(extent(0,2,0,2), res = 1, vals = c(1,1,2,2))
+  a <- raster(extent(0, 2, 0, 2), res = 1, vals = c(1, 1, 2, 2))
   levels(a) <- data.frame(ID = 1:2, Factor = c("This", "That"))
-  tifTmp <- tempfile(fileext=".tif")
-  b1 <- writeRaster(a, filename = tifTmp)
+  tifTmp <- tempfile(fileext = ".tif")
+  file.create(tifTmp)
+  tifTmp <- normPath(tifTmp)
+
+  b1 <- writeRaster(a, filename = tifTmp, overwrite = TRUE)
   expect_warning(b1a <- writeOutputs(a, filename2 = tifTmp))
   expect_false(identical(b1, b1a))
   expect_true(identical(as.integer(b1[]), b1a[]))
+
   expect_true(identical(filename(b1), tifTmp))
   expect_true(identical(filename(b1a), gsub(tifTmp, pattern = "tif", replacement = "grd")))
-
 })