+++ NEWS.md
@@ -13,6 +13,7 @@ version 0.2.4
 * files in `CHECKSUMS` now sorted alphabetically
 * `Checksums` can now have a `CHECKSUMS.txt` file located in a different place than the `destinationPath`
 * added checks for float point number issues in raster resolutions produced by `raster::projectRaster`
+* `.robustDigest` now does not include `Cache`-added attributes
 * minor bugfixes
 * many new unit tests written, which caught several minor bugs